requirements.txt
Flask
Flask-cors
python-dotenv
uwsgi
requests